Extract EntryAtRevision from DraftEntry and PublishedEntry
Reduce duplication between classes.

REDMINE-18710
